Refactor history and settings pages for Next.js compliance
